Added Python script replacement for pipeline
This is just an early draft. It gets as far as trimming then
joining the reads, followed by conversion to FASTA. Some changes
still to be done:

The application/script wrappers look like they could be heavily
refactored to inherit from a single class.

Still to do the QC, and later stages of the pipeline.